+++ src/operators/SurfaceNormal/avtSurfaceNormalFilter.C
@@ -158,7 +158,7 @@ avtSurfaceNormalFilter::InferVariableNameFromContract(avtContract_p c)
     if (!foundIt)
     {
         std::vector<CharStrRef>   vars2nd = in_dr->GetSecondaryVariablesWithoutDuplicates();
-        for (int i = 0 ; i < vars2nd.size() ; i++)
+        for (size_t i = 0 ; i < vars2nd.size() ; i++)
         if (strncmp(*(vars2nd[i]), "operators/SurfaceNormal/", 
                     strlen("operators/SurfaceNormal/")) == 0)
         {
